backend/core/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==21.2.0
whitenoise==6.7.0

